ENSIBS-4/eduroom/server/app-container/requirements.txt
fastapi==0.86.0
uvicorn==0.19.0
pyJWT==2.5.0
python-decouple==3.6
email-validator==1.3.0
itsdangerous==2.1.2
python-cas==1.6.0
sqlalchemy==1.4.42
cryptography==38.0.3
urllib3==1.26.3
mysql-connector-python==8.0.31
bcrypt==4.0.1
fastapi-mail
Jinja2==3.1.2
openpyxl==3.0.10
requests==2.25.1
starlette==0.20.4
python-multipart==0.0.5